requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
Automat==20.2.0
boto3==1.17.48
botocore==1.20.48
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
DateTime==4.3
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.2
django-admin==2.0.1
django-channels==0.7.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-messages==0.6.0
django-six==1.0.4
django-storages==1.11.1
django-templated-mail==1.1.1
django-tempus-dominus==5.1.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
drf-nested-routers==0.93.4
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
jsonpickle==1.4.1
Markdown==3.3.4
MarkupSafe==2.0.1
mechanize==0.4.5
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.3.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.6
screen==1.0.1
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
tqdm==4.52.0
txaio==20.12.1
ujson==3.2.0
uritemplate==4.1.1
urllib3==1.26.2
virtualenv==20.0.33
webencodings==0.5.1
whitenoise==5.2.0
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.2.0
